+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.6.0
 
 - generate the full load tree and use `.load(...).execute(...)` to avoid exponential N+1 problem during deletion cascades
+- types now allow `optional` option for DreamSerializer `delegatedAttribute` (used by Psychic to customize OpenAPI shape of automatically inferred fields to allow `null`)
 
 ## 2.5.8